(refactor) : 암호화 기능별 구분 #EAT-38
