--- src/main/java/team/gwon/haveameal/member/encryptionservice/EncryptionService.java
@@ -1,9 +1,0 @@
-package team.gwon.haveameal.member.encryptionservice;
-
-public interface EncryptionService {
-	String encryptPassword(String plainPassword);
-
-	String encryptInfo(String plainInfo);
-
-	String decryptInfo(String encryptionData);
-}
