+++ src/main/java/team/gwon/haveameal/member/encryptionservice/password/BCryptPasswordEncryptor.java
@@ -1,0 +1,12 @@
+package team.gwon.haveameal.member.encryptionservice.password;
+
+import org.springframework.security.crypto.bcrypt.BCrypt;
+import org.springframework.stereotype.Component;
+
+@Component
+public class BCryptPasswordEncryptor implements PasswordEncryptor {
+	@Override
+	public String encryptPassword(String plainPassword) {
+		return BCrypt.hashpw(plainPassword, BCrypt.gensalt());
+	}
+}
