+++ src/main/java/team/gwon/haveameal/member/encryptionservice/password/PasswordEncryptor.java
@@ -1,0 +1,5 @@
+package team.gwon.haveameal.member.encryptionservice.password;
+
+public interface PasswordEncryptor {
+	String encryptPassword(String plainPassword);
+}
