+++ src/main/java/team/gwon/haveameal/member/encryptionservice/personaldata/PersonalDataDecryptor.java
@@ -1,0 +1,5 @@
+package team.gwon.haveameal.member.encryptionservice.personaldata;
+
+public interface PersonalDataDecryptor {
+	String decryptData(String encryptionData);
+}
