+++ src/main/java/team/gwon/haveameal/member/encryptionservice/personaldata/PersonalDataEncryptor.java
@@ -1,0 +1,5 @@
+package team.gwon.haveameal.member.encryptionservice.personaldata;
+
+public interface PersonalDataEncryptor {
+	String encryptData(String plainData);
+}
